+++ qim3d/viz/__init__.py
@@ -1,6 +1,8 @@
 from .visualizations import plot_metrics
-from .img import grid_pred, grid_overview, slices, slicer, orthogonal, plot_cc, local_thickness
+from .img import grid_pred, grid_overview, slices, slicer, orthogonal
 from .k3d import vol
 from .structure_tensor import vectors
+from .local_thickness_ import local_thickness
+from .cc import plot_cc
 from .colormaps import objects
 from .detection import circles