+++ qim3d/viz/img.py
@@ -3,7 +3,7 @@ Provides a collection of visualization functions.
 """
 
 import math
-from typing import List, Optional, Union, Tuple
+from typing import List, Optional, Union
 
 import ipywidgets as widgets
 import matplotlib.pyplot as plt
@@ -12,10 +12,8 @@ import torch
 from matplotlib import colormaps
 from matplotlib.colors import LinearSegmentedColormap
 
-import qim3d.io
 from qim3d.io.logger import log
-from qim3d.utils.cc import CC
-from qim3d.viz.colormaps import objects
+
 
 
 def grid_overview(
@@ -270,7 +268,7 @@ def slices(
         img = qim3d.examples.shell_225x128x128
         qim3d.viz.slices(img, n_slices=15)
         ```
-
+        ![Grid of slices](assets/screenshots/viz-slices.png)
     """
 
     # Numpy array or Torch tensor input
@@ -411,8 +409,8 @@ def slicer(
         vol (np.ndarray or torch.Tensor): The 3D volume to be sliced.
         axis (int, optional): Specifies the axis, or dimension, along which to slice. Defaults to 0.
         cmap (str, optional): Specifies the color map for the image. Defaults to "viridis".
-        img_height(int, optional): Height of the figure. Defaults to 3.
-        img_width(int, optional): Width of the figure. Defaults to 3.
+        img_height (int, optional): Height of the figure. Defaults to 3.
+        img_width (int, optional): Width of the figure. Defaults to 3.
         show_position (bool, optional): If True, displays the position of the slices. Defaults to False.
         interpolation (str, optional): Specifies the interpolation method for the image. Defaults to None.
 
@@ -424,8 +422,9 @@ def slicer(
         import qim3d
 
         vol = qim3d.examples.bone_128x128x128
-        qim3d.viz.slicer(vol, cmap="magma")
+        qim3d.viz.slicer(vol)
         ```
+        ![viz slicer](assets/screenshots/viz-slicer.gif)
     """
 
     # Create the interactive widget
@@ -483,9 +482,10 @@ def orthogonal(
         ```python
         import qim3d
 
-        vol = qim3d.examples.bone_128x128x128
-        qim3d.viz.orthogonal(vol)
+        vol = qim3d.examples.fly_150x256x256
+        qim3d.viz.orthogonal(vol, cmap="magma")
         ```
+        ![viz orthogonal](assets/screenshots/viz-orthogonal.gif)
     """
 
     z_slicer = slicer(
@@ -521,193 +521,3 @@ def orthogonal(
     x_slicer.children[0].description = "X"
 
     return widgets.HBox([z_slicer, y_slicer, x_slicer])
-
-
-def plot_cc(
-    connected_components: CC,
-    component_indexs: list | tuple = None,
-    max_cc_to_plot=32,
-    overlay=None,
-    crop=False,
-    show=True,
-    **kwargs,
-) -> list[plt.Figure]:
-    """
-    Plot the connected components of an image.
-
-    Parameters:
-        connected_components (CC): The connected components object.
-        components (list | tuple, optional): The components to plot. If None the first max_cc_to_plot=32 components will be plotted. Defaults to None.
-        max_cc_to_plot (int, optional): The maximum number of connected components to plot. Defaults to 32.
-        overlay (optional): Overlay image. Defaults to None.
-        crop (bool, optional): Whether to crop the image to the cc. Defaults to False.
-        show (bool, optional): Whether to show the figure. Defaults to True.
-        **kwargs: Additional keyword arguments to pass to `qim3d.viz.slices`.
-
-    Returns:
-        figs (list[plt.Figure]): List of figures, if `show=False`.
-    """
-    # if no components are given, plot the first max_cc_to_plot=32 components
-    if component_indexs is None:
-        if len(connected_components) > max_cc_to_plot:
-            log.warning(
-                f"More than {max_cc_to_plot} connected components found. Only the first {max_cc_to_plot} will be plotted. Change max_cc_to_plot to plot more components."
-            )
-        component_indexs = range(
-            1, min(max_cc_to_plot + 1, len(connected_components) + 1)
-        )
-        
-    figs = []
-    for component in component_indexs:
-        if overlay is not None:
-            assert (overlay.shape == connected_components.shape), f"Overlay image must have the same shape as the connected components. overlay.shape=={overlay.shape} != connected_components.shape={connected_components.shape}."
-
-            # plots overlay masked to connected component
-            if crop:
-                # Crop the overlay image based on the bounding box of the component
-                bb = connected_components.get_bounding_box(component)[0]
-                cc = connected_components.get_cc(component, crop=True)
-                overlay_crop = overlay[bb]
-                # use cc as mask for overlay_crop, where all values in cc set to 0 should be masked out, cc contains integers
-                overlay_crop = np.where(cc == 0, 0, overlay_crop)
-                fig = slices(overlay_crop, show=show, **kwargs)
-            else:
-                cc = connected_components.get_cc(component, crop=False)
-                overlay_crop = np.where(cc == 0, 0, overlay)
-                fig = slices(overlay_crop, show=show, **kwargs)
-        else:
-            # assigns discrete color map to each connected component if not given 
-            if "cmap" not in kwargs:
-                kwargs["cmap"] = qim3dCmap(len(component_indexs))
-        
-            # Plot the connected component without overlay
-            fig = slices(connected_components.get_cc(component, crop=crop), show=show, **kwargs)
-
-        figs.append(fig)
-
-    if not show:
-        return figs
-
-    return
-
-
-def local_thickness(
-    image: np.ndarray,
-    image_lt: np.ndarray,
-    max_projection: bool = False,
-    axis: int = 0,
-    slice_idx: Optional[Union[int, float]] = None,
-    show: bool = False,
-    figsize: Tuple[int, int] = (15, 5),
-) -> Union[plt.Figure, widgets.interactive]:
-    """Visualizes the local thickness of a 2D or 3D image.
-
-    Args:
-        image (np.ndarray): 2D or 3D NumPy array representing the image/volume.
-        image_lt (np.ndarray): 2D or 3D NumPy array representing the local thickness of the input
-            image/volume.
-        max_projection (bool, optional): If True, displays the maximum projection of the local
-            thickness. Only used for 3D images. Defaults to False.
-        axis (int, optional): The axis along which to visualize the local thickness.
-            Unused for 2D images.
-            Defaults to 0.
-        slice_idx (int or float, optional): The initial slice to be visualized. The slice index
-            can afterwards be changed. If value is an integer, it will be the index of the slice
-            to be visualized. If value is a float between 0 and 1, it will be multiplied by the
-            number of slices and rounded to the nearest integer. If None, the middle slice will
-            be used for 3D images. Unused for 2D images. Defaults to None.
-        show (bool, optional): If True, displays the plot (i.e. calls plt.show()). Defaults to False.
-        figsize (Tuple[int, int], optional): The size of the figure. Defaults to (15, 5).
-
-    Raises:
-        ValueError: If the slice index is not an integer or a float between 0 and 1.
-
-    Returns:
-        If the input is 3D, returns an interactive widget. Otherwise, returns a matplotlib figure.
-
-    Example:
-        image_lt = qim3d.processing.local_thickness(image)
-        qim3d.viz.local_thickness(image, image_lt, slice_idx=10)
-    """
-
-    def _local_thickness(image, image_lt, show, figsize, axis=None, slice_idx=None):
-        if slice_idx is not None:
-            image = image.take(slice_idx, axis=axis)
-            image_lt = image_lt.take(slice_idx, axis=axis)
-
-        fig, axs = plt.subplots(1, 3, figsize=figsize, layout="constrained")
-
-        axs[0].imshow(image, cmap="gray")
-        axs[0].set_title("Original image")
-        axs[0].axis("off")
-
-        axs[1].imshow(image_lt, cmap="viridis")
-        axs[1].set_title("Local thickness")
-        axs[1].axis("off")
-
-        plt.colorbar(
-            axs[1].imshow(image_lt, cmap="viridis"), ax=axs[1], orientation="vertical"
-        )
-
-        axs[2].hist(image_lt[image_lt > 0].ravel(), bins=32, edgecolor="black")
-        axs[2].set_title("Local thickness histogram")
-        axs[2].set_xlabel("Local thickness")
-        axs[2].set_ylabel("Count")
-
-        if show:
-            plt.show()
-
-        plt.close()
-
-        return fig
-
-    # Get the middle slice if the input is 3D
-    if len(image.shape) == 3:
-        if max_projection:
-            if slice_idx is not None:
-                log.warning(
-                    "slice_idx is not used for max_projection. It will be ignored."
-                )
-            image = image.max(axis=axis)
-            image_lt = image_lt.max(axis=axis)
-            return _local_thickness(image, image_lt, show, figsize)
-        else:
-            if slice_idx is None:
-                slice_idx = image.shape[axis] // 2
-            elif isinstance(slice_idx, float):
-                if slice_idx < 0 or slice_idx > 1:
-                    raise ValueError(
-                        "Values of slice_idx of float type must be between 0 and 1."
-                    )
-                slice_idx = int(slice_idx * image.shape[0]) - 1
-            slide_idx_slider = widgets.IntSlider(
-                min=0,
-                max=image.shape[axis] - 1,
-                step=1,
-                value=slice_idx,
-                description="Slice index",
-                layout=widgets.Layout(width="450px"),
-            )
-            widget_obj = widgets.interactive(
-                _local_thickness,
-                image=widgets.fixed(image),
-                image_lt=widgets.fixed(image_lt),
-                show=widgets.fixed(True),
-                figsize=widgets.fixed(figsize),
-                axis=widgets.fixed(axis),
-                slice_idx=slide_idx_slider,
-            )
-            widget_obj.layout = widgets.Layout(align_items="center")
-            if show:
-                display(widget_obj)
-            return widget_obj
-    else:
-        if max_projection:
-            log.warning(
-                "max_projection is only used for 3D images. It will be ignored."
-            )
-        if slice_idx is not None:
-            log.warning("slice_idx is only used for 3D images. It will be ignored.")
-        return _local_thickness(image, image_lt, show, figsize)
-
-